+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,36 @@
+import * as path from 'path';
+import {release, version} from 'os';
+import { createServer as createServerHttp } from 'node:http'
+import './files/c.js';
+
+import { fileURLToPath } from 'node:url';
+import { dirname } from 'node:path';
+
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = await import('./files/a.json', { assert: { type: 'json' } });
+} else {
+    unknownObject = await import('./files/b.json', { assert: { type: 'json' } });
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const createMyServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+export {
+    unknownObject,
+    createMyServer,
+};
